优化了webui添加好友的部分，以及修复了主动回复消息无法设置小时的问题

+++ src/config/__init__.py
@@ -43,8 +43,8 @@ class MediaSettings:
 @dataclass
 class AutoMessageSettings:
     content: str
-    min_hours: int
-    max_hours: int
+    min_hours: float
+    max_hours: float
 
 @dataclass
 class QuietTimeSettings:
